deleted one StreamFragment, fixed presenter methods, code cleanup
