+++ app/src/main/java/ru/gorshenev/themesstyles/presentation/ui/channels/StreamFragment.kt
@@ -1,13 +1,14 @@
 package ru.gorshenev.themesstyles.presentation.ui.channels
 
 import android.os.Bundle
+import android.util.Log
 import android.view.View
 import androidx.core.os.bundleOf
 import androidx.fragment.app.Fragment
 import by.kirich1409.viewbindingdelegate.viewBinding
 import com.google.android.material.snackbar.Snackbar
 import ru.gorshenev.themesstyles.R
-import ru.gorshenev.themesstyles.data.Utils.setDivider
+import ru.gorshenev.themesstyles.utils.Utils.setDivider
 import ru.gorshenev.themesstyles.databinding.FragmentChannelsStreamBinding
 import ru.gorshenev.themesstyles.presentation.base_recycler_view.Adapter
 import ru.gorshenev.themesstyles.presentation.base_recycler_view.HolderFactory
@@ -15,27 +16,29 @@ import ru.gorshenev.themesstyles.presentation.base_recycler_view.ViewTyped
 import ru.gorshenev.themesstyles.presentation.ui.channels.ChannelsFragment.Companion.RESULT_STREAM
 import ru.gorshenev.themesstyles.presentation.ui.channels.ChannelsFragment.Companion.STREAM_SEARCH
 import ru.gorshenev.themesstyles.presentation.ui.channels.ChannelsFragment.Companion.STR_NAME
+import ru.gorshenev.themesstyles.presentation.ui.channels.ChannelsFragment.Companion.STR_TYPE
 import ru.gorshenev.themesstyles.presentation.ui.channels.ChannelsFragment.Companion.TPC_NAME
 import ru.gorshenev.themesstyles.presentation.ui.channels.adapter.StreamsHolderFactory
 import ru.gorshenev.themesstyles.presentation.ui.channels.items.StreamUi
 import ru.gorshenev.themesstyles.presentation.ui.channels.items.TopicUi
 import ru.gorshenev.themesstyles.presentation.ui.chat.ChatFragment
 
-class StreamSubsFragment : Fragment(R.layout.fragment_channels_stream), StreamView {
+class StreamFragment : Fragment(R.layout.fragment_channels_stream), StreamView {
     private val binding: FragmentChannelsStreamBinding by viewBinding()
 
     private val presenter: StreamPresenter = StreamPresenter(this)
 
     private val holderFactory: HolderFactory = StreamsHolderFactory(
         onStreamClick = { streamId -> presenter.onStreamClick(streamId) },
-        onTopicClick = { topicId -> onTopicClick(topicId) }
+        onTopicClick = { topicId -> presenter.onTopicClick(topicId) }
     )
     private val adapter = Adapter<ViewTyped>(holderFactory)
 
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         initViews()
-        presenter.loadStreams(30)
+        val streamType = arguments?.get(STR_TYPE) as StreamType
+        presenter.loadStreams(30, streamType)
     }
 
     override fun onDestroyView() {
@@ -44,14 +47,21 @@ class StreamSubsFragment : Fragment(R.layout.fragment_channels_stream), StreamVi
     }
 
 
-    private fun onTopicClick(topicId: Int) {
-        val topic = adapter.items.find { (it is TopicUi) && it.id == topicId }
-        val stream = adapter.items.find { (it is StreamUi) && it.topics.contains(topic) }
+    private fun initViews() {
+        with(binding) {
+            rvStreams.adapter = adapter
+            rvStreams.setDivider()
+
+            parentFragmentManager.setFragmentResultListener(STREAM_SEARCH, this@StreamFragment) { _, result ->
+                val searchText = result.getString(RESULT_STREAM, "")
+                presenter.searchStream(searchText?.toString().orEmpty())
+            }
+        }
+    }
+
+    override fun goToChat(topic: TopicUi, stream: StreamUi) {
         val chatFragment = ChatFragment()
-        chatFragment.arguments = bundleOf(
-            STR_NAME to (stream as StreamUi).name,
-            TPC_NAME to (topic as TopicUi).name
-        )
+        chatFragment.arguments = bundleOf(STR_NAME to stream.name, TPC_NAME to topic.name)
 
         parentFragmentManager.beginTransaction()
             .add(R.id.fragment_container_view, chatFragment)
@@ -59,38 +69,22 @@ class StreamSubsFragment : Fragment(R.layout.fragment_channels_stream), StreamVi
             .commit()
     }
 
-    private fun initViews() {
-        with(binding) {
-            rvStreams.adapter = adapter
-            rvStreams.setDivider()
-
-            parentFragmentManager.setFragmentResultListener(
-                STREAM_SEARCH,
-                this@StreamSubsFragment
-            ) { _, result ->
-                val searchText = result.getString(RESULT_STREAM, "")
-                presenter.searchStream(searchText?.toString().orEmpty())
-            }
-        }
-    }
-
     override fun showError(error: Throwable?) {
         Snackbar.make(binding.root, "Something wrong! $error", Snackbar.LENGTH_SHORT).show()
     }
 
-    override fun adapterItems(): List<ViewTyped> {
-        return adapter.items
-    }
-
     override fun showLoading() {
-        TODO("Not yet implemented")
+        binding.shimmerChannels.apply {
+            visibility = View.VISIBLE
+            showShimmer(true)
+        }
     }
 
     override fun stopLoading() {
         Snackbar.make(binding.root, "Completed", Snackbar.LENGTH_SHORT).show()
         binding.shimmerChannels.apply {
             visibility = View.GONE
-            stopShimmer()
+            hideShimmer()
         }
     }
 
@@ -98,5 +92,9 @@ class StreamSubsFragment : Fragment(R.layout.fragment_channels_stream), StreamVi
         adapter.items = items
     }
 
+     enum class StreamType {
+        SUBSCRIBED,
+        ALL_STREAMS
+    }
 
 }
