+++ app/src/main/java/ru/gorshenev/themesstyles/presentation/ui/channels/StreamView.kt
@@ -1,6 +1,8 @@
 package ru.gorshenev.themesstyles.presentation.ui.channels
 
 import ru.gorshenev.themesstyles.presentation.base_recycler_view.ViewTyped
+import ru.gorshenev.themesstyles.presentation.ui.channels.items.StreamUi
+import ru.gorshenev.themesstyles.presentation.ui.channels.items.TopicUi
 
 interface StreamView {
 
@@ -12,6 +14,6 @@ interface StreamView {
 
     fun showError(error: Throwable?)
 
-    fun adapterItems(): List<ViewTyped>
+    fun goToChat(topic: TopicUi, stream: StreamUi)
 
 }
