+++ app/src/main/java/ru/gorshenev/themesstyles/presentation/ui/chat/adapter/MessageRightViewHolder.kt
@@ -3,8 +3,8 @@ package ru.gorshenev.themesstyles.presentation.ui.chat.adapter
 import android.view.View
 import by.kirich1409.viewbindingdelegate.viewBinding
 import ru.gorshenev.themesstyles.presentation.ui.chat.views.EmojiView
-import ru.gorshenev.themesstyles.data.Utils.px
-import ru.gorshenev.themesstyles.data.Utils.toEmojiString
+import ru.gorshenev.themesstyles.utils.Utils.px
+import ru.gorshenev.themesstyles.utils.Utils.toEmojiString
 import ru.gorshenev.themesstyles.presentation.base_recycler_view.BaseViewHolder
 import ru.gorshenev.themesstyles.databinding.ViewCustomViewGroupRightBinding
 import ru.gorshenev.themesstyles.presentation.ui.chat.items.MessageRightUi
