+++ app/src/main/java/ru/gorshenev/themesstyles/presentation/ui/chat/adapter/ReactionsViewHolder.kt
@@ -3,7 +3,7 @@ package ru.gorshenev.themesstyles.presentation.ui.chat.adapter
 import android.view.View
 import by.kirich1409.viewbindingdelegate.viewBinding
 import ru.gorshenev.themesstyles.presentation.base_recycler_view.BaseViewHolder
-import ru.gorshenev.themesstyles.data.Utils.toEmojiString
+import ru.gorshenev.themesstyles.utils.Utils.toEmojiString
 import ru.gorshenev.themesstyles.databinding.ItemEmojiBinding
 import ru.gorshenev.themesstyles.presentation.ui.chat.items.ReactionsUi
