+++ app/src/main/java/ru/gorshenev/themesstyles/presentation/ui/people/PeoplePresenter.kt
@@ -1,12 +1,14 @@
 package ru.gorshenev.themesstyles.presentation.ui.people
 
+import io.reactivex.Single
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
 import io.reactivex.subjects.PublishSubject
-import ru.gorshenev.themesstyles.data.Utils
+import ru.gorshenev.themesstyles.utils.Utils
 import ru.gorshenev.themesstyles.data.repositories.PeopleDataSource
 import ru.gorshenev.themesstyles.presentation.base_recycler_view.ViewTyped
+import ru.gorshenev.themesstyles.presentation.ui.people.items.PeopleUi
 import java.util.concurrent.TimeUnit
 
 class PeoplePresenter(private val view: PeopleView) {
@@ -17,12 +19,11 @@ class PeoplePresenter(private val view: PeopleView) {
 
     private val searchSubject: PublishSubject<String> = PublishSubject.create()
 
-    //todo почему?
     init {
         searchSubject
             .distinctUntilChanged()
             .debounce(500, TimeUnit.MILLISECONDS)
-            .switchMapSingle { searchTxt -> Utils.initUserSearch(cachedItems, searchTxt) }
+            .switchMapSingle { searchTxt -> initUserSearch(cachedItems, searchTxt) }
             .subscribeOn(Schedulers.computation())
             .observeOn(AndroidSchedulers.mainThread())
             .subscribe(
@@ -33,11 +34,10 @@ class PeoplePresenter(private val view: PeopleView) {
     }
 
     fun loadPeople(count: Int) {
-        view.showLoading()
         PeopleDataSource.getPeople(count)
             .subscribeOn(Schedulers.io())
-            .debounce(500, TimeUnit.MILLISECONDS)
             .observeOn(AndroidSchedulers.mainThread())
+            .doOnSubscribe { view.showLoading() }
             .subscribe(
                 { people ->
                     view.showItems(people)
@@ -54,6 +54,20 @@ class PeoplePresenter(private val view: PeopleView) {
 
     fun onClear() {
         compositeDisposable.clear()
+    }
+
+    private fun initUserSearch(
+        cachedItems: List<ViewTyped>,
+        searchText: String
+    ): Single<List<ViewTyped>> {
+        return Single.fromCallable {
+            return@fromCallable when {
+                searchText.isEmpty() -> cachedItems
+                else -> cachedItems.filter { item ->
+                    item is PeopleUi && (item.name.contains(searchText, true) || item.email.contains(searchText, true))
+                }
+            }
+        }
     }
 
 }
